% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularity.R
\name{compute_modularity}
\alias{compute_modularity}
\title{Modularity}
\usage{
compute_modularity(g, method = igraph::cluster_louvain, ...)
}
\arguments{
\item{g}{a graph. If the graph is not connected, the largest connected
component is selected.}

\item{method}{default is \code{\link[igraph]{cluster_louvain}}, but others
are possible: e.g. \code{\link[igraph]{cluster_fast_greedy}},
\code{\link[igraph]{cluster_infomap}}.}

\item{...}{additional parameters passed to the chosen igraph clustering
method. The default \code{\link[igraph]{cluster_louvain}} has a
\code{weights} parameter, ehich can be a positive weight vector, or the
\code{E(g)$weight} attribute, or \code{NA} if the graph has a 'weight'
edge attribute, but you want to ignore it.
Larger edge weights correspond to stronger connections.
Check out parameters for others methods.}
}
\value{
list(Q, N) where:
\itemize{
  \item{Q}{the modularity}
  \item{N}{the number of communities}
}
}
\description{
Compute the modularity and number communities of graph g,
   w.r.t. its community structure, obtained through the chosen clustering
   method.
}
\references{
Newman
}
\seealso{
[igraph::cluster_infomap()]
}
\keyword{modularity;}
\keyword{networks}
\keyword{segregation;}
