% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCE.R
\name{GCE}
\alias{GCE}
\title{Global Communication Efficeiency}
\usage{
GCE(g, directed = FALSE, normalised = TRUE, weights = NULL, verbose = TRUE)
}
\arguments{
\item{g}{a network.}

\item{directed}{logical, if the directed network has to be considered.
If FALSE (default) the network is taken as undirected.}

\item{normalised}{logical, default TRUE.}

\item{weights}{edge weights, representing the appeal of each link.
if weights is NULL (the default) and g has a weight edge
attribute that will be passed to the igraph
\code{\link[igraph]{distances}} function, as \eqn{w_{ij}^{-1}}.}

\item{verbose}{logical, default TRUE.}
}
\value{
A list
  \itemize{
    \item non_normalised - communication efficiency \eqn{E(G)},
    \item normalised - normalised communication efficiency, \eqn{GCE(G)},
      or NULL if normalised is \code{FALSE}.
  }
}
\description{
The average inverse shortest path length is a measure known as
   the global efficiency (see Latora and Marchiori, 2001).
   We implement the global communication efficiency (GCE) for weighted
   networks and propose a new normalisation method for the GCE.
   The global efficiency may be meaningfully computed on disconnected
   networks, as paths between disconnected nodes are defined to have
   infinite length and correspondingly zero efficiency.
   \strong{Assumption}: Edge weights are non-negative and represent the
   \emph{appeal} of links, hence we want shortest paths with heavy links.

   \strong{Non-normalised GCE}:
   \deqn{E(g) = \frac{1}{N} \sum_{i \in V} \frac{\sum_{j \in V,i \neq j} d_{ij}}{N - 1},}
   where \eqn{N} is the number of vertices of graph \eqn{g} and \eqn{d_{ij}}
   is the shortest-path distance between \eqn{i} and \eqn{j}, computed by
   the Floyd-Warshall's algorithm with inverse edge weights as edge costs.
   See \url{https://en.wikipedia.org/wiki/Floyd-Warshall_algorithm}.
   \strong{Normalisation}: the previous efficiency should be divided by the
   efficency of an \emph{ideal} graph.
   \deqn{E(g_{ideal}) = \frac{1}{N} \sum_{i \in V} \frac{\sum_{j \in V,i \neq j} l_{ij}}{N - 1}.}
   The normalised communication efficiency is then given by
   \deqn{GCE(g)=\frac{E(g)}{E(g_{ideal})}}
}
\examples{
library(intsegration)
library(igraph)
karate <- make_graph("zachary")
GCE(karate, directed = F)
}
\references{
\describe{
    \item{[1]}{Latora, V. & Marchiori, M. (2001).
               Efficient Behavior of Small-World Networks.
               \url{https://doi.org/10.1103/PhysRevLett.87.198701}}
    \item{[2]}{Bertagnolli, G., Gallotti R. & De Domenico, M. (2020)
               Quantifying efficient information exchange in real flow
               networks. \url{arxiv}}
   }
}
\keyword{communication}
\keyword{efficiency;}
\keyword{integration;}
