% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularity.R
\name{LCC_subgraph}
\alias{LCC_subgraph}
\title{Isolate the Largest Connected Component}
\usage{
LCC_subgraph(g)
}
\arguments{
\item{g}{a graph.}
}
\value{
list(LCC, leftovers) where:

* LCC is the largest connected components
* leftovers the nodes (labels or indeces depending on the network) that do
not belong to the LCC.
}
\description{
Compute modularity and number communities
}
