% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCE.R
\name{LCE}
\alias{LCE}
\title{Local Communication Efficeiency}
\usage{
LCE(g, directed = FALSE, normalised = TRUE, weights = NULL)
}
\arguments{
\item{g}{a network.}

\item{directed}{default FALSE.}

\item{normalised}{logical, default TRUE.}

\item{weights}{edge weights, representing the appeal of each link.
if weights is NULL (the default) and g has a weight edge
attribute that will be passed to the igraph
\code{\link[igraph]{distances}} function, as \eqn{w_{ij}^{-1}}.}
}
\value{
local communication efficiency vector, for each node \eqn{i \in G}
   \eqn{LCE(G_i)} if \code{normalised = TRUE}, \eqn{E(G_i)} otherwise.
}
\description{
The average inverse shortest path length is a measure known as
   the global efficiency (see Latora and Marchiori, 2001).
   We implement the local communication efficiency of node \eqn{i \in G} as
   the global communication efficency of the subgraph induced by \eqn{i},
   \code{\link{GCE}}.
}
\examples{
library(intsegration)
library(igraph)
karate <- make_graph("zachary")
LCE(karate)
}
\references{
Latora, V., & Marchiori, M. (2001). Efficient Behavior of Small-World Networks. \url{https://doi.org/10.1103/PhysRevLett.87.198701}
}
\keyword{centrality;}
\keyword{communication}
\keyword{efficiency;}
\keyword{integration;}
