% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_floyd_flow}
\alias{rcpp_floyd_flow}
\title{Floyd-Warshall algorithm for shortest-path distances and total flows}
\usage{
rcpp_floyd_flow(C)
}
\arguments{
\item{C}{numeric matrix of edge costs}
}
\value{
D matrix of shortest-path distances, i.e. sum of costs along minimising paths

F matrix of total flows, i.e. sum of inverse costs, along shortest-paths
}
\description{
Floyd-Warshall algorithm for shortest-path distances and total flows
}
