% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_floyd_flow_length}
\alias{rcpp_floyd_flow_length}
\title{Floyd-Warshall algorithm for shortest-path distances and total flows}
\usage{
rcpp_floyd_flow_length(C, L)
}
\arguments{
\item{C}{numeric matrix of edge costs}

\item{L}{numeric matrix of edge (Euclidean) lenghts}
}
\value{
D matrix of shortest-path distances, i.e. sum of costs along minimising paths

F matrix of total flows, i.e. sum of inverse costs, along shortest-paths

D_eucl matrix of Euclidean distances along shortest-paths (paths minimising sum of costs)

F_eucl matrix of total Eucl. flows, i.e. sum of inverse lengths, along shortest-paths
}
\description{
Additionally, it computes also the Euclidean length of shortest-paths (sum of Euclidean lengths)
and the total Euclidean closeness (sum of reciprocals Euclidean lengths) along shortest-paths.
}
